Replace bulk_save_objects() with add_all()
For some reason, bulk_save_objects() was causing an extremely strange error in which "COREQUISITE" and "CROSSLIST" were attempted to be used as values in the course_relationship table even though I was using enum values in the Python code.

My only guess as to why is because bulk_save_objects "stringifies" objects without resolving variable names and somehow interpreted models.RelationshipTypeEnum.COREQUISITE as the string "COREQUISITE", somehow.